+++ tests/integ/modin/series/test_dt_accessor.py
@@ -93,7 +93,7 @@ def test_date(datetime_index_value):
 @pytest.mark.parametrize(
     "datetime_index_value",
     [
-        ["2014-04-04 23:56", "2014-07-18 21:24", "2015-11-22 22:14"],
+        ["2014-04-04 23:56:20", "2014-07-18 21:24:30", "2015-11-22 22:14:40"],
         ["04/04/2014", "07/18/2013", "11/22/2015"],
         ["2014-04-04 23:56", pd.NaT, "2014-07-18 21:24", "2015-11-22 22:14", pd.NaT],
         [
@@ -103,7 +103,7 @@ def test_date(datetime_index_value):
         ],
     ],
 )
-@pytest.mark.parametrize("func", ["floor", "ceil"])
+@pytest.mark.parametrize("func", ["floor", "ceil", "round"])
 @pytest.mark.parametrize(
     "freq",
     [
@@ -117,16 +117,21 @@ def test_date(datetime_index_value):
         "2s",
     ],
 )
-@sql_count_checker(query_count=1)
-def test_floor_ceil(datetime_index_value, func, freq):
+def test_floor_ceil_round(datetime_index_value, func, freq):
     native_ser = native_pd.Series(native_pd.DatetimeIndex(datetime_index_value))
     snow_ser = pd.Series(native_ser)
-    eval_snowpark_pandas_result(
-        snow_ser, native_ser, lambda ser: getattr(ser.dt, func)(freq)
-    )
+    if func == "round" and "s" in freq:
+        with SqlCounter(query_count=0):
+            with pytest.raises(NotImplementedError):
+                snow_ser.dt.round(freq=freq)
+    else:
+        with SqlCounter(query_count=1):
+            eval_snowpark_pandas_result(
+                snow_ser, native_ser, lambda ser: getattr(ser.dt, func)(freq)
+            )
 
 
-@pytest.mark.parametrize("func", ["floor", "ceil"])
+@pytest.mark.parametrize("func", ["floor", "ceil", "round"])
 @pytest.mark.parametrize(
     "freq, ambiguous, nonexistent",
     [
@@ -137,7 +142,7 @@ def test_floor_ceil(datetime_index_value, func, freq):
     ],
 )
 @sql_count_checker(query_count=0)
-def test_floor_ceil_negative(func, freq, ambiguous, nonexistent):
+def test_floor_ceil_round_negative(func, freq, ambiguous, nonexistent):
     datetime_index_value = [
         "2014-04-04 23:56",
         pd.NaT,